+++ main-service/main-server/src/main/java/ru/practicum/event/Event.java
@@ -18,20 +18,20 @@ import java.time.LocalDateTime;
 public class Event {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int id;
+    long id;
     String annotation;
-    int categoryId;
-    int confirmedRequests;
+    long categoryId;
+    long confirmedRequests;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     LocalDateTime createdOn;
     String description;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     LocalDateTime eventDate;
-    int initiatorId;
+    long initiatorId;
     float location_lat;
     float location_lon;
     Boolean paid;
-    int participantLimit;
+    long participantLimit;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     LocalDateTime publishedOn;
     Boolean requestModeration;
